+++ examples/text_layouter_inline_boxes.rb
@@ -41,16 +41,14 @@ size = 10
 items = sample_text.split(/(:-\)|;-\))/).map do |part|
   case part
   when ':-)'
-    style = {background_color: [162, 234, 247], padding: 2}
-    InlineBox.create(content_width: size * 2, content_height: size * 2,
-                     style: style) do |canvas, box|
+    InlineBox.create(width: size * 2, height: size * 2, content_box: true,
+                     background_color: [162, 234, 247], padding: 2) do |canvas, box|
       canvas.image(emoji_smile, at: [0, 0], width: box.content_width)
     end
   when ';-)'
-    style = {padding: 5, margin: [0, 10],
-             border: {width: [1, 2], color: [200, 40]}}
-    InlineBox.create(content_width: size, content_height: size,
-                     valign: :top, style: style) do |canvas, box|
+    InlineBox.create(width: size, height: size, content_box: true,
+                     valign: :top, padding: 5, margin: [0, 10],
+                     border: {width: [1, 2], color: [200, 40]}) do |canvas, box|
       canvas.image(emoji_wink, at: [0, 0], width: box.content_width)
     end
   else